requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Data Processing
pandas==2.0.3

# Database
duckdb==0.9.2

# LLM Integration
huggingface-hub==0.20.3
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
